transport/ardop: Refactor and cleanup
