+++ transport/ardop/tnc.go
@@ -6,7 +6,6 @@ package ardop
 
 import (
 	"bufio"
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"io"
@@ -135,86 +134,6 @@ func (tnc *TNC) init() (err error) {
 
 var ErrChecksumMismatch = fmt.Errorf("Control protocol checksum mismatch")
 
-func readFrame(reader *bufio.Reader) (frame, error) {
-	fType, err := reader.ReadByte()
-	if err != nil {
-		return nil, err
-	}
-
-	reader.Discard(1) // :
-
-	var data []byte
-	switch fType {
-	case 'c':
-		data, err = reader.ReadBytes('\r')
-	case 'd':
-		// Peek length
-		peeked, err := reader.Peek(2)
-		if err != nil {
-			return nil, err
-		}
-		length := binary.BigEndian.Uint16(peeked) + 2 // +2 to include the length bytes
-
-		// actual data
-		data = make([]byte, length)
-		var n int
-		for read := 0; read < int(length) && err == nil; {
-			n, err = reader.Read(data[read:])
-			read += n
-		}
-	default:
-		return nil, fmt.Errorf("Unexpected frame type %c", fType)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	// Verify CRC sums
-	sumBytes := make([]byte, 2)
-	reader.Read(sumBytes)
-	crc := binary.BigEndian.Uint16(sumBytes)
-	if crc16Sum(data) != crc {
-		return nil, ErrChecksumMismatch
-	}
-
-	switch fType {
-	case 'c':
-		data = data[:len(data)-1] // Trim \r
-		return cmdFrame(string(data)), nil
-	case 'd':
-		return dFrame{dataType: string(data[2:5]), data: data[5:]}, nil
-	default:
-		panic("not possible")
-	}
-}
-
-func writeLine(w io.Writer, format string, params ...interface{}) error {
-	payload := fmt.Sprintf(format+"\r", params...)
-	if _, err := fmt.Fprint(w, "C:"+payload); err != nil {
-		return err
-	}
-
-	sum := crc16Sum([]byte(payload))
-	return binary.Write(w, binary.BigEndian, sum)
-}
-
-type frame interface{}
-
-type dFrame struct {
-	dataType string
-	data     []byte
-}
-
-func (f dFrame) ARQFrame() bool { return f.dataType == "ARQ" }
-func (f dFrame) FECFrame() bool { return f.dataType == "FEC" }
-func (f dFrame) ErrFrame() bool { return f.dataType == "ERR" }
-func (f dFrame) IDFFrame() bool { return f.dataType == "IDF" }
-
-type cmdFrame string
-
-func (f cmdFrame) Parsed() ctrlMsg { return parseCtrlMsg(string(f)) }
-
 func (tnc *TNC) runControlLoop() error {
 	rd := bufio.NewReader(tnc.ctrl)
 
@@ -319,7 +238,7 @@ func (tnc *TNC) runControlLoop() error {
 					log.Println("-->", str)
 				}
 
-				if err := writeLine(tnc.ctrl, str); err != nil {
+				if err := writeCtrlFrame(tnc.ctrl, str); err != nil {
 					panic(err)
 				}
 			case data, ok := <-dataOut:
